修改抽獎錯誤及餐廳首頁按鈕

+++ homeworks/week17/hw2/migrations/20210827093409-create-prize.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       prizeProbability: {
-        type: Sequelize.DECIMAL(10, 2)
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
